requirements_ppo.txt
accelerate==1.1.1
certifi==2024.8.30
charset-normalizer==3.4.0
click==8.1.7
docker-pycreds==0.4.0
filelock==3.16.1
fsspec==2024.10.0
gitdb==4.0.11
GitPython==3.1.43
huggingface-hub==0.26.2
idna==3.10
Jinja2==3.1.4
MarkupSafe==3.0.2
mpmath==1.3.0
numpy==1.26.4
packaging==24.2
platformdirs==4.3.6
protobuf==5.28.3
psutil==6.1.0
PyYAML==6.0.2
regex==2024.11.6
requests==2.32.3
safetensors==0.4.5
sentry-sdk==2.18.0
setproctitle==1.3.4
six==1.16.0
smmap==5.0.1
sympy==1.13.3
tokenizers==0.20.3
torch==2.2.2
tqdm==4.67.0
transformers==4.40.2
typing_extensions==4.12.2
urllib3==2.2.3
wandb==0.18.7
peft==0.7.0
trl==0.9.4